requirements.txt
django
python-dotenv
psycopg2
djangorestframework
django-filter